rename file to align with cli name
